更新GDTSDK

+++ libGDTMobSDK/GDTSplashAd.h
@@ -85,7 +85,7 @@
  *  详解：拉取广告超时时间，开发者调用loadAd方法以后会立即展示backgroundColor，然后在该超时时间内，如果广告拉
  *  取成功，则立马展示开屏广告，否则放弃此次广告展示机会。
  */
-@property (nonatomic, assign) int fetchDelay;
+@property (nonatomic, assign) NSInteger fetchDelay;
 
 /**
  *  开屏广告的背景色
@@ -96,7 +96,7 @@
 /**
  * 跳过按钮的位置
  */
-@property (nonatomic, assign) CGPoint  skipButtonCenter;
+@property (nonatomic, assign) CGPoint skipButtonCenter;
 
 /**
  *  构造方法
